requirements.txt
ansible>=2.9.0
PyYAML>=5.1
Jinja2>=2.10
qrcode[pil]>=6.1
cryptography>=3.0
requests>=2.20
